integration fix test packages after ArchUnit refactoring
