+++ digiwf-integrations/digiwf-address-integration/digiwf-address-integration-core/src/test/java/de/muenchen/oss/digiwf/address/integration/application/usecase/AddressesGermanyUseCaseTest.java
@@ -1,8 +1,7 @@
-package integration.application.usecase;
+package de.muenchen.oss.digiwf.address.integration.application.usecase;
 
 import de.muenchen.oss.digiwf.address.integration.application.port.in.AddressGermanyInPort;
 import de.muenchen.oss.digiwf.address.integration.application.port.out.AddressClientOutPort;
-import de.muenchen.oss.digiwf.address.integration.application.usecase.AddressesGermanyUseCase;
 import de.muenchen.oss.digiwf.address.integration.client.gen.model.BundesweiteAdresseResponse;
 import de.muenchen.oss.digiwf.address.integration.client.model.request.SearchAddressesGermanyModel;
 import de.muenchen.oss.digiwf.message.process.api.error.BpmnError;
