+++ digiwf-integrations/digiwf-address-integration/digiwf-address-integration-core/src/test/java/de/muenchen/oss/digiwf/address/integration/application/usecase/StreetsMunichUseCaseTest.java
@@ -1,8 +1,7 @@
-package integration.application.usecase;
+package de.muenchen.oss.digiwf.address.integration.application.usecase;
 
 import de.muenchen.oss.digiwf.address.integration.application.port.in.StreetsMunichInPort;
 import de.muenchen.oss.digiwf.address.integration.application.port.out.AddressClientOutPort;
-import de.muenchen.oss.digiwf.address.integration.application.usecase.StreetsMunichUseCase;
 import de.muenchen.oss.digiwf.address.integration.client.gen.model.Strasse;
 import de.muenchen.oss.digiwf.address.integration.client.gen.model.StrasseResponse;
 import de.muenchen.oss.digiwf.address.integration.client.model.request.ListStreetsModel;
