+++ digiwf-integrations/digiwf-dms-integration/digiwf-dms-integration-core/src/test/java/de/muenchen/oss/digiwf/dms/integration/adapter/in/streaming/DepositObjectStreamingAdapterTest.java
@@ -17,7 +17,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-class DepositObjectStreamingAdapterTest extends MessageProcessorTestBase {
+class DepositObjectStreamingAdapterTest extends StreamingAdapterTestBase {
 
     private final DepositObjectDto depositObjectDto = new DepositObjectDto(
             "objectCoo",
